Move to SearchHit
